+++ Assets/Scripts/Player/Skills/Player Actions/PoisonData.cs
@@ -19,7 +19,7 @@ public class PoisonData : PlayerAction {
         m_skillData = PlayerSkillManager.Instance.GetPlayerSkillData(PLAYER_SKILL_TYPE.POISON);
         m_playerSkillData = PlayerSkillManager.Instance.GetPlayerSkillData<PlayerSkillData>(PLAYER_SKILL_TYPE.POISON);
         //IncreaseThreatForEveryCharacterThatSeesPOI(targetPOI, 5);
-        targetPOI.traitContainer.AddTrait(targetPOI, "Poisoned", overrideDuration: (int)m_playerSkillData.skillUpgradeData.GetDurationBonusPerLevel(m_skillData.currentLevel));
+        targetPOI.traitContainer.AddTrait(targetPOI, "Poisoned", overrideDuration: m_playerSkillData.skillUpgradeData.GetDurationBonusPerLevel(m_skillData.currentLevel));
         Log log = GameManager.CreateNewLog(GameManager.Instance.Today(), "InterventionAbility", name, "activated", null, LOG_TAG.Player);
         log.AddToFillers(targetPOI, targetPOI.name, LOG_IDENTIFIER.ACTIVE_CHARACTER);
         log.AddLogToDatabase();
